requirements.txt
asgiref==3.5.0
boto3==1.21.44
botocore==1.24.44
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
Django==4.0.4
django-cognito-jwt==0.0.4
djangorestframework==3.13.1
drf-yasg==1.20.0
ecdsa==0.17.0
envs==1.4
future==0.18.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
packaging==21.3
pyasn1==0.4.8
pycognito==2022.3.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.3.1
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2022.1
requests==2.27.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
uWSGI==2.0.20
